refacto: move global font from AppNavigator.js to App.js
